Refactor lambdas to be one unit instead of smaller one.
