+++ userservice/mapping-templates/getUsers-request.vtl
@@ -2,6 +2,7 @@
     "version": "2017-02-28",
     "operation": "Invoke",
     "payload": {
+        "fieldName" : "GetUsers",
         "input":  $utils.toJson($context.arguments)
     }
 }
