+++ userservice/serverless.yml
@@ -31,14 +31,8 @@ provider:
           - Arn
 
 functions:
-  GetUsers:
-    handler: src/lambda/getUsers.handler
-  AddUser:
-    handler: src/lambda/addUser.handler
-  UpdateUser:
-    handler: src/lambda/updateUser.handler
-  DeleteUser:
-    handler: src/lambda/deleteUser.handler
+  userService:
+    handler: src/lambda/UserServiceHandler.handler
   graphql:
     handler: src/graphql/graphql.graphqlHandler
     events:
@@ -56,47 +50,32 @@ custom:
     name: ${self:service}_${self:provider.stage}
     authenticationType: API_KEY
     mappingTemplates:
-      - dataSource: getUsers
+      - dataSource: UserService
         type: Query
         field: getUsers
         request: "getUsers-request.vtl"
         response: "getUsers-response.vtl"
-      - dataSource: addUser
+      - dataSource: UserService
         type: Mutation
         field: addUser
         request: "addUser-request.vtl"
         response: "addUser-response.vtl"
-      - dataSource: updateUser
+      - dataSource: UserService
         type: Mutation
         field: updateUser
         request: "updateUser-request.vtl"
         response: "updateUser-response.vtl"
-      - dataSource: deleteUser
+      - dataSource: UserService
         type: Mutation
         field: deleteUser
         request: "deleteUser-request.vtl"
         response: "deleteUser-response.vtl"
     dataSources:
       - type: AWS_LAMBDA
-        name: getUsers
-        description: 'Lambda getUsers DataSource'
+        name: UserService
+        description: 'Lambda UserService DataSource'
         config:
-          functionName: GetUsers
-      - type: AWS_LAMBDA
-        name: addUser
-        description: 'Lambda addUser DataSource'
-        config:
-          functionName: AddUser
-      - type: AWS_LAMBDA
-        name: updateUser
-        description: 'Lambda updateUser DataSource'
-        config:
-          functionName: UpdateUser
-      - type: AWS_LAMBDA
-        name: deleteUser
-        description: 'Lambda deleteUser DataSource'
-        config:
-          functionName: DeleteUser
+          functionName: userService
   dynamodb:
     stages:
       - test
