--- userservice/src/lambda/getUsers.js
@@ -1,16 +1,0 @@
-import getUserService from "../compositeRoot/getUserService.js";
-import logError from "../error/logError.js";
-
-exports.handler = async (event) => {
-  try {
-    const userService = getUserService();
-    return await userService.getUsers(
-      event.input.filter,
-      event.input.limit,
-      event.input.startKey
-    );
-  } catch (e) {
-    logError(e);
-    throw e;
-  }
-};
