+++ userservice/src/utils/constants.js
@@ -1,0 +1,9 @@
+const Constants = {
+  ADD_USER: "AddUser",
+  GET_USER: "GetUser",
+  GET_USERS: "GetUsers",
+  UPDATE_USER: "UpdateUser",
+  DELETE_USER: "DeleteUser",
+};
+
+export default Constants;
