--- userservice/test/integration/lambda/getUsers.test.js
@@ -1,115 +1,0 @@
-import "../EnvironmentVariable.js";
-import "@jest/globals";
-import { afterAll, expect } from "@jest/globals";
-
-import { handler } from "../../../src/lambda/getUsers.js";
-import { handler as addUserHandler } from "../../../src/lambda/addUser.js";
-import { handler as deleteUserHandler } from "../../../src/lambda/deleteUser.js";
-
-describe("getUsers", () => {
-  let seedItems = [];
-  beforeAll(async () => {
-    jest.setTimeout(30000);
-    for (let i = 0; i < 10; i++) {
-      let item = await addUserHandler({
-        input: {
-          name: `user ${i}`,
-          dateOfBirth: new Date().valueOf(),
-          address: `address ${i}`,
-          description: `description ${i}`,
-        },
-      });
-      seedItems.push(item);
-    }
-
-    let item = await addUserHandler({
-      input: {
-        name: `2 users`,
-        dateOfBirth: new Date().valueOf(),
-        address: ``,
-        description: ``,
-      },
-    });
-    seedItems.push(item);
-
-    item = await addUserHandler({
-      input: {
-        name: `2 users`,
-        dateOfBirth: new Date().valueOf(),
-        address: ``,
-        description: ``,
-      },
-    });
-    seedItems.push(item);
-  });
-
-  afterAll(async () => {
-    for (let i = 0; i < seedItems.length; i++) {
-      await deleteUserHandler({
-        input: seedItems[i],
-      });
-    }
-  });
-
-  test(`Input with a limit of 6, 6 users return`, async () => {
-    let result = await handler({
-      input: {
-        filter: null,
-        limit: 6,
-        startKey: null,
-      },
-    });
-
-    expect(result.items.length).toBe(6);
-    expect(result.items[5].id).toBe(seedItems[6].id);
-  });
-
-  test(`Input with a limit of 1 and a start key seedItem in index 7, User at index 6 return`, async () => {
-    let result = await handler({
-      input: {
-        filter: null,
-        limit: 1,
-        startKey: { id: seedItems[7].id },
-      },
-    });
-
-    expect(result.items.length).toBe(1);
-    expect(result.items[0].id).toBe(seedItems[6].id);
-  });
-
-  test(`Input with a limit of 6 and filter with the name "2 users", 2 users return`, async () => {
-    let result = await handler({
-      input: {
-        filter: "2 users",
-        limit: 6,
-        startKey: null,
-      },
-    });
-
-    expect(result.items.length).toBe(2);
-  });
-
-  test(`Input with a limit of 6 and filter with the name "2 uSeRs", 2 users return`, async () => {
-    let result = await handler({
-      input: {
-        filter: "2 uSeRs",
-        limit: 6,
-        startKey: null,
-      },
-    });
-
-    expect(result.items.length).toBe(2);
-  });
-
-  test(`Input with a limit of 100 and filter with the name that does not exist, 0 user return`, async () => {
-    let result = await handler({
-      input: {
-        filter: "that does not exist",
-        limit: 100,
-        startKey: null,
-      },
-    });
-
-    expect(result.items.length).toBe(0);
-  });
-});
